requirements.txt
# requirements.txt
langchain
langchain-openai
streamlit
pandas
plotly
duckdb
python-dotenv
openpyxl
pydantic